♻️  Refactor: move updateEvent to event repo
