+++ packages/backend/src/event/services/recur/repo/event.repo.ts
@@ -1,8 +1,12 @@
 /**
  * Event DB operations repo for Compass's Event collection
  */
-import { ObjectId } from "mongodb";
-import { Event_Core } from "@core/types/event.types";
+import { ObjectId, WithId } from "mongodb";
+import {
+  Event_Core,
+  Schema_Event,
+  Schema_Event_Core,
+} from "@core/types/event.types";
 import { Collections } from "@backend/common/constants/collections";
 import mongoService from "@backend/common/services/mongo.service";
 import { Ids_Event } from "@backend/event/queries/event.queries";
@@ -35,5 +39,33 @@ export class EventRepository {
         { upsert: true },
       );
     return result;
+  }
+
+  /**
+   * Replace an event document by _id and user. Throws if no match.
+   */
+  async updateEvent(eventId: string, event: Schema_Event_Core) {
+    const _event = {
+      ...event,
+      user: this.userId,
+    };
+    if ("_id" in event) {
+      // WARNING: This is a hack to avoid the _id field from being updated
+      // This may result in unexpected behavior
+      delete _event._id;
+    }
+    const response = (await mongoService.db
+      .collection(Collections.EVENT)
+      .findOneAndReplace(
+        { _id: new ObjectId(eventId), user: this.userId },
+        _event,
+        { returnDocument: "after" },
+      )) as unknown as WithId<Schema_Event>;
+    if (!response) {
+      throw new Error(
+        "Event not updated due to failed DB operation (NoMatchingEvent)",
+      );
+    }
+    return response;
   }
 }
